Add support for multiple cleaners per column
Lets the user apply as many functions to a column as desired through the
addition of a list of transformations passed to column.transform instead
of a single transformer.  ORDER DOES MATTER! The order in which the
transforms are listed is the order in which they will be applied.  It's
best to keep transforms as idempotent as possible to minimize the impact
of cleaner order on the application.